requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pydantic
slowapi
transformers
torch
pytesseract
pdf2image
pillow
